Refactor Crawl Scheduler controllers to centralize and  enhance functionality.
Removed `CrawlSchedulerStatus` controller and merged its functionality into `CrawlSchedulerUtilsController` to streamline code. Updated endpoint mappings for consistency and applied security annotations where needed in the backend. No frontend changes were involved.
